Moved default authenticator into internal package. Fixed error in default authenticator where it tried to register the same http handler multiple times. Added config decorator functions WithDefaultOAUTH, WithClientCredentials Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>
