+++ internal/auth/oauth/defauth/html_template.go
@@ -1,4 +1,4 @@
-package dev
+package defauth
 
 import (
 	"bytes"
