+++ internal/auth/oauth/defauth/templates/simple.html
