+++ internal/client/client.go
@@ -22,6 +22,7 @@ import (
 	"github.com/apache/thrift/lib/go/thrift"
 	"github.com/databricks/databricks-sql-go/auth"
 	"github.com/databricks/databricks-sql-go/driverctx"
+	"github.com/databricks/databricks-sql-go/internal/auth/oauth/defauth"
 	"github.com/databricks/databricks-sql-go/internal/cli_service"
 	"github.com/databricks/databricks-sql-go/internal/config"
 	"github.com/databricks/databricks-sql-go/logger"
@@ -253,7 +254,10 @@ func InitThriftClient(cfg *config.Config, httpclient *http.Client) (*ThriftServi
 	case "http":
 		if httpclient == nil {
 			if cfg.Authenticator == nil {
-				return nil, dbsqlerrint.NewRequestError(context.TODO(), dbsqlerr.ErrNoAuthenticationMethod, nil)
+				cfg.Authenticator, err = defauth.NewDefaultAuthenticator(cfg.Host, 0)
+				if err != nil {
+					return nil, dbsqlerrint.NewRequestError(context.TODO(), dbsqlerr.ErrNoDefaultAuthenticator, nil)
+				}
 			}
 			httpclient = RetryableClient(cfg)
 		}
@@ -403,7 +407,11 @@ func PooledTransport() *http.Transport {
 
 func PooledClient(cfg *config.Config) *http.Client {
 	if cfg.Authenticator == nil {
-		return nil
+		var err error
+		cfg.Authenticator, err = defauth.NewDefaultAuthenticator(cfg.Host, 0)
+		if err != nil {
+			return nil
+		}
 	}
 
 	var tr *Transport
